requirements.txt
python-hcl2==4.3.2
scikit-learn==1.3.2
numpy==1.24.3
joblib==1.3.2
coverage==7.3.2
pytest==7.4.3
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
prometheus-client==0.19.0
pyyaml==6.0.1
